Deprecate the `NR_TAGS` and `NR_ENV_DELIMITER` configuration options
The new `ADDITIONAL_ATTRIBUTES` provides a full replacement, so the logic now is
to use what is configured in that variable, and only if that is unset fall back
to parsing `NR_TAGS` and `NR_ENV_DELIMITER`.

Note that `ADDITIONAL_ATTRIBUTES` is a JSON string, and doesn't contain the logic
to not add `aws:`- and `plugin:`-prefixed properties: If those aren't wanted they
shouldn't be provided.